// src/types/supabase.ts
export type Database = {
  __InternalSupabase: {
    PostgrestVersion: "12.2.3 (519615d)"
  }
  public: {
    Tables: {
      income: {
        Row: {
          base_salary: number
          created_at: string | null
          housing_fund: number | null
          housing_fund_deduction: number | null
          id: string
          leave_deduction: number | null
          medical_insurance: number | null
          overtime_meal: number | null
          owner: string
          pension_insurance: number | null
          rent: number | null
          tax: number | null
          time: string
          unemployment_insurance: number | null
          updated_at: string | null
        }
        Insert: {
          base_salary: number
          created_at?: string | null
          housing_fund?: number | null
          housing_fund_deduction?: number | null
          id?: string
          leave_deduction?: number | null
          medical_insurance?: number | null
          overtime_meal?: number | null
          owner: string
          pension_insurance?: number | null
          rent?: number | null
          tax?: number | null
          time: string
          unemployment_insurance?: number | null
          updated_at?: string | null
        }
        Update: {
          base_salary?: number
          created_at?: string | null
          housing_fund?: number | null
          housing_fund_deduction?: number | null
          id?: string
          leave_deduction?: number | null
          medical_insurance?: number | null
          overtime_meal?: number | null
          owner?: string
          pension_insurance?: number | null
          rent?: number | null
          tax?: number | null
          time?: string
          unemployment_insurance?: number | null
          updated_at?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never